requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Cors==4.0.0
Flask-SocketIO==5.3.6
python-dotenv==1.0.0
passlib==1.7.4
bcrypt==4.1.2
gunicorn==21.2.0
requests==2.31.0
itsdangerous==2.1.2
email-validator==2.3.1
SQLAlchemy==2.0.23
alembic==1.13.0
psycopg2-binary==2.9.9
Flask-Migrate==4.0.7
pandas==2.1.4
numpy==1.26.2
scikit-learn==1.3.2
matplotlib==3.8.2
seaborn==0.13.0
flask-jwt-extended==4.6.0
Flask-Bcrypt==0.7.1
python-jose==3.3.0
validators==0.22.0
pillow==10.1.0
protobuf==4.25.1
google-cloud-storage==2.14.0
google-auth==2.23.4
google-cloud-core==2.3.2
opencv-python==4.8.1.78
firebase-admin==6.3.0
Flask-Session==0.4.0
Werkzeug==3.0.1
markupsafe==2.1.3
greenlet==3.0.1
jinja2==3.1.3
sqlalchemy-utils==0.41.1
cachelib==0.8.1
click==8.1.7
Flask-APScheduler==1.12.3
grequests==0.7.0
gevent==23.9.1
requests-oauthlib==1.3.1
oauthlib==3.2.2
twilio==8.7.0
stripe==7.7.0
paypalrestsdk==1.13.0
Flask-Mail==0.9.1
celery==5.3.6
redis==5.0.1
flower==2.0.1
eventlet==0.34.0
pika==1.3.2
sentry-sdk==1.39.1
qrcode==7.4.2
imagecode==0.3.0
reportlab==4.0.7
zbarlight==1.3
pyzbar==0.2.0
colorama==0.4.6
pytest==7.4.3
pytest-cov==4.1.0
coverage==7.3.2
flake8==6.1.0
mypy==1.6.1
isort==5.12.0
black==23.10.0
autopep8==2.0.4
bandit==1.7.5
safety==3.0.0
pycodestyle==2.11.1
Flask-Testing==0.8.1
factory-boy==3.3.0
Faker==19.13.0
deepdiff==6.7.1
pytest-flask==1.3.0
requests-mock==1.11.0
mock==5.1.0
freezegun==1.2.2
respx==0.20.1
asgi-lifespan==2.1.0
httpx==0.25.0
nest-asyncio==1.5.8
pytest-asyncio==0.23.2
types-requests==2.31.0.10
chardet==5.2.0
certifi==2023.11.17
idna==3.6
urllib3==1.26.18
typing-extensions==4.8.0
toml==0.10.2
attrs==23.2.0
click-plugins==1.1.1
pygments==2.16.1
mako==1.3.0
packaging==23.2
markupsafe==2.1.3
sqlalchemy==2.0.23
typing-extensions==4.8.0
nest-asyncio==1.5.8
pytest-asyncio==0.23.2
types-requests==2.31.0.10
toml==0.10.2
attrs==23.2.0
click-plugins==1.1.1
pygments==2.16.1
mako==1.3.0
packaging==23.2
python-dotenv==1.0.0
passlib==1.7.4
bcrypt==4.1.2
gunicorn==21.2.0
requests==2.31.0
itsdangerous==2.1.2
email-validator==2.3.1
sqlalchemy==2.0.23
alembic==1.13.0
psycopg2-binary==2.9.9
flask-migrate==4.0.7
pandas==2.1.4
numpy==1.26.2
scikit-learn==1.3.2
matplotlib==3.8.2
seaborn==0.13.0
flask-jwt-extended==4.6.0
flask-bcrypt==0.7.1
python-jose==3.3.0
validators==0.22.0
pillow==10.1.0
protobuf==4.25.1
google-cloud-storage==2.14.0
google-auth==2.23.4
google-cloud-core==2.3.2
opencv-python==4.8.1.78
firebase-admin==6.3.0
flask-session==0.4.0
werkzeug==3.0.1
markupsafe==2.1.3
greenlet==3.0.1
jinja2==3.1.3
sqlalchemy-utils==0.41.1
cachelib==0.8.1
click==8.1.7
flask-apscheduler==1.12.3
grequests==0.7.0
gevent==23.9.1
requests-oauthlib==1.3.1
oauthlib==3.2.2
twilio==8.7.0
stripe==7.7.0
paypalrestsdk==1.13.0
flask-mail==0.9.1
celery==5.3.6
redis==5.0.1
flower==2.0.1
eventlet==0.34.0
pika==1.3.2
sentry-sdk==1.39.1
qrcode==7.4.2
imagecode==0.3.0
reportlab==4.0.7
zbarlight==1.3
pyzbar==0.2.0
colorama==0.4.6
pytest==7.4.3
pytest-cov==4.1.0
coverage==7.3.2
flake8==6.1.0
mypy==1.6.1
isort==5.12.0
black==23.10.0
autopep8==2.0.4
bandit==1.7.5
safety==3.0.0
pycodestyle==2.11.1
flask-testing==0.8.1
factory-boy==3.3.0
faker==19.13.0
deepdiff==6.7.1
pytest-flask==1.3.0
requests-mock==1.11.0
mock==5.1.0
freezegun==1.2.2
respx==0.20.1
asgi-lifespan==2.1.0
httpx==0.25.0
nest-asyncio==1.5.8
pytest-asyncio==0.23.2
types-requests==2.31.0.10
chardet==5.2.0
certifi==2023.11.17
idna==3.6
urllib3==1.26.18
typing-extensions==4.8.0
toml==0.10.2
attrs==23.2.0
click-plugins==1.1.1
pygments==2.16.1
mako==1.3.0
packaging==23.2
markupsafe==2.1.3
sqlalchemy==2.0.23
typing-extensions==4.8.0
nest-asyncio==1.5.8
pytest-asyncio==0.23.2
types-requests==2.31.0.10
toml==0.10.2
attrs==23.2.0
click-plugins==1.1.1
pygments==2.16.1
mako==1.3.0
packaging==23.2
chardet==5.2.0
certifi==2023.11.17
idna==3.6
urllib3==1.26.18
typing-extensions==4.8.0
toml==0.10.2
attrs==23.2.0
click-plugins==1.1.1
pygments==2.16.1
mako==1.3.0
packaging==23.2
markupsafe==2.1.3
sqlalchemy==2.0.23
typing-extensions==4.8.0
nest-asyncio==1.5.8
pytest-asyncio==0.23.2
types-requests==2.31.0.10
toml==0.10.2
attrs==23.2.0
click-plugins==1.1.1
pygments==2.16.1
mako==1.3.0
packaging==23.2